+++ cores/domain/src/main/java/com/theloungemembers/core/api/repository/ApiUsageLogJpaRepository.java
@@ -5,5 +5,5 @@ import com.theloungemembers.core.api.entity.ApiUsageLogEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface ApiUsageLogJpaRepository extends JpaRepository<ApiUsageLogEntity, Integer>, JpaSpecificationExecutor<ApiUsageLogEntity> {
+public interface ApiUsageLogJpaRepository extends JpaRepository<ApiUsageLogEntity, Long>, JpaSpecificationExecutor<ApiUsageLogEntity> {
 }
