+++ cores/model/src/main/java/com/theloungemembers/core/api/ApiAccessibleIpCommand.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Setter
 public class ApiAccessibleIpCommand {
 
-    private Integer uid;
+    private Long uid;
 
     private String ipAddress;
